reduce time before end of test game

+++ Sword.cpp
@@ -14,9 +14,9 @@ Sword::Sword(bool isTwoHanded) :
 	}
 	// ONE HAND SWORD
 	else {
-		m_offHandSkill = new Skill(m_baseDamage * getoffHandsDamageMultiplicator(), m_baseRange);
+		m_offHandSkill = new Skill(static_cast<int>(m_baseDamage * getoffHandsDamageMultiplicator()), m_baseRange);
 		m_offHandSkill->setName("SAIGNE CHAROGNE !");
-		m_offHandSkill->addEffect(Effect::EffectCategory::PRIMARY, new Effect(Effect::EffectType::DEBUFF, Effect::EffectTarget::HEALTH, Effect::SkillType::STRENGTH, 15*getoffHandsDamageMultiplicator(), 3, false));
+		m_offHandSkill->addEffect(Effect::EffectCategory::PRIMARY, new Effect(Effect::EffectType::DEBUFF, Effect::EffectTarget::HEALTH, Effect::SkillType::STRENGTH, static_cast<int>(15*getoffHandsDamageMultiplicator()), 3, false));
 		m_offHandSkill->addEffect(Effect::EffectCategory::SECONDARY, new Effect(Effect::EffectType::DEBUFF, Effect::EffectTarget::MOVEMENT, Effect::SkillType::DEXTERITY, 1, 3, true));
 	}
 	m_skills.push_back(m_mainSkill);
